Switch from properties to context-param
Turns out that envVars can’t be set from the pom for jetty:run, values
in app.yam can’t have filters on them, and since we are in war, it’s
harder to substitute in properties files. (since they aren’t auto
loaded). So, we switch to web.xml which allows filters (if you enable
it).

1. Switch to context-param, and change code to use it.
2. add vars to pom, so they can be changed w/ -D on the mvn line.
3. Capitalize comments where appropriate.
4. Add a logging directive to app.yam
5. Tell .gitignore to ignore intellij.