Hide image upload when no Cloud Storage config.
We have developers run the sample without doing any configuration. It's
not a great experience if the app seems broken when they create a book
and try to include an image.

I only hide the element since the Create and Update servlets expect the
"file" property to be populated (even if it's empty string).

We should consider doing the same for the login button and "my books"
button if the client ID for OAuth2 is not set.