[GAE Std J8 getting started] - Datastore sample
Part of a new Getting Started experience for GAE Standard Java 8. This sample shows how to setup a simple J8 app - a blog - using Cloud Datastore as the storage backend.

Notes:
It is intended to be a standalone folder, hence the pom.xml does not have a parent.
Region tags for documentation to be added later.